# sigpcb-base-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Base",
        "description": "Gestion des API de base",
        "contact": {
            "name": "Support client",
            "email": "contact@exemple.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/anatt-base/centre-examens": {
            "get": {
                "tags": [
                    "CentreExamens"
                ],
                "summary": "Récupérer la liste des centres d'examen",
                "description": "Récupère une liste de tous les centres d'examen enregistrés dans la base de données",
                "operationId": "getAllCentreExamens",
                "responses": {
                    "200": {
                        "description": "La liste des centres d'examen récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID du centre",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "description": "Nom du centre",
                                                "type": "string"
                                            },
                                            "commune_id": {
                                                "description": "ID de la commune à laquelle appartient le nouveau centre créé",
                                                "type": "integer"
                                            },
                                            "status": {
                                                "description": "Statut du centre (optionnel)",
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CentreExamens"
                ],
                "summary": "Crée un nouveau centre d'examen",
                "description": "Crée un nouveau centre d'examen enregistré dans la base de données",
                "operationId": "createCentreExamen",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom du centre d'examen",
                                        "type": "string"
                                    },
                                    "commune_id": {
                                        "description": "ID de la commune à laquelle appartient le nouveau centre créé",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "Statut du centre (optionnel)",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouveau centre d'examen créé"
                    }
                }
            }
        },
        "/api/anatt-base/centre-examens/{id}": {
            "get": {
                "tags": [
                    "CentreExamens"
                ],
                "summary": "Récupère un centre par ID",
                "description": "Récupère un centre enregistré dans la base de données en spécifiant son ID",
                "operationId": "getCentreExamenById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du centre à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centre d'examen récupéré avec succès"
                    },
                    "404": {
                        "description": "Centre d'examen non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "CentreExamens"
                ],
                "summary": "Met à jour un centre existant",
                "description": "Met à jour un centre existant dans la base de données",
                "operationId": "updateCentreExamen",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du centre à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom du centre d'examen",
                                        "type": "string"
                                    },
                                    "commune_id": {
                                        "description": "ID de la commune à laquelle appartient le nouveau centre créé",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "Statut du centre (optionnel)",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Centre d'examen mis à jour avec succès"
                    },
                    "404": {
                        "description": "Centre d'examen non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CentreExamens"
                ],
                "summary": "Supprime un centre",
                "description": "Supprime un centre de la base de données",
                "operationId": "deleteCentreExamen",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du centre à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centre d'examen supprimé avec succès"
                    },
                    "404": {
                        "description": "Centre d'examen non trouvé"
                    }
                }
            }
        },
...
    }
}